refactor(logic): split deal_matcher into part-of concern files (Refs #3290)
Phase 0 file decomposition for the colonizethis_logic package split.
deal_matcher.dart (598 physical lines) exceeded the >500-line Phase 0
target, so its private static helpers are extracted into two `part of`
concern files that share the library's private scope:

- deal_matcher_matching.dart: FRR pass, lock-recovery seller ordering,
  integer-priority tier matching.
- deal_matcher_indexing.dart: order indexing/aggregation, carry-forward,
  and per-buyer treasury/affordability helpers.

The public DealMatcher.matchDeals / DealMatcher.pairKey API and the
DealMatchInputs typedef are unchanged. Behaviour-preserving: helpers move
to top-level private functions in the same library, so call resolution and
the deterministic matching semantics are identical (the single cross-class
reference, pairKey inside tier matching, is now qualified as
DealMatcher.pairKey). Each resulting file is well under the repo file-size
policy (262 / 215 / 147 physical lines).
