+++ packages/colonizethis_logic/lib/src/economy/world_market/deal_matcher.dart
@@ -16,11 +16,21 @@
 /// `originTileKey` that resolves through [PurchasedTileIndex], the owning
 /// Great Power's bids for the same commodity match against that offer
 /// before any integer-priority tier or FTP pair runs.
+///
+/// The matching passes and the order-indexing / treasury-affordability
+/// helpers live in `part of` concern files
+/// (`deal_matcher_matching.dart`, `deal_matcher_indexing.dart`) to keep
+/// each file below the repo file-size policy (`SPEC/program/
+/// dart-file-non-comment-line-size.md`); they share this library's private
+/// scope so behaviour is unchanged (Refs #3290 Phase 0 file decomposition).
 library;
 
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import 'purchased_tile_index.dart';
+
+part 'deal_matcher_matching.dart';
+part 'deal_matcher_indexing.dart';
 
 /// Inputs for a single [DealMatcher.matchDeals] pass.
 ///
@@ -248,351 +258,5 @@ class DealMatcher {
       unfilledBidsByFactionId: _carryForwardByFaction(bidStatesByFaction),
       activityByCommodityId: Map.unmodifiable(activity),
     );
-  }
-
-  /// Runs the First Right of Refusal absolute-priority pass for one
-  /// commodity per `SPEC/game/world-market-first-right-of-refusal.md` and
-  /// `SPEC/program/world-market-resolution.md` § Step B (absolute-priority
-  /// tier above tier 1).
-  ///
-  /// For each offer whose `originTileKey` resolves through
-  /// [purchasedTileIndex], iterate the owning Great Power's bids for the
-  /// same commodity in their submitted (`factionLocalIndex`) order and
-  /// match them ahead of any integer-priority tier. FTP membership is
-  /// ignored — FRR overrides FTP per `SPEC/game/world-market.md`.
-  /// `FilledDeal.isFirstRightOfRefusalMatch` is set to true on emitted
-  /// deals so D4 treasury transfers and the Deal Book UI can identify
-  /// FRR-applied flows. Cargo is consumed normally per buyer.
-  ///
-  /// Offers iterate in `(sellerFactionId, factionLocalIndex)` order to
-  /// preserve determinism. Bids iterate in `factionLocalIndex` order
-  /// within the owning GP's bid list.
-  static int _runFirstRightMatching({
-    required List<_OrderState> commodityOffers,
-    required List<_OrderState> commodityBids,
-    required CommodityId commodityId,
-    required double pricePerUnit,
-    required PurchasedTileIndex purchasedTileIndex,
-    required Map<String, int> remainingCargo,
-    required Map<String, int> remainingTreasury,
-    required List<FilledDeal> filledOut,
-  }) {
-    if (commodityOffers.isEmpty || commodityBids.isEmpty) return 0;
-    var filledQuantity = 0;
-
-    int attemptFrrMatch(_OrderState offer, _OrderState bid) {
-      if (offer.remaining <= 0 || bid.remaining <= 0) return 0;
-      final cargoLeft = remainingCargo[bid.factionId] ?? 0;
-      if (cargoLeft <= 0) return 0;
-      final desiredQty = _min3(offer.remaining, bid.remaining, cargoLeft);
-      if (desiredQty <= 0) return 0;
-      final maxAffordable = _maxAffordableQuantity(
-        bid: bid,
-        pricePerUnit: pricePerUnit,
-        remainingTreasury: remainingTreasury,
-      );
-      final matchQty = desiredQty <= maxAffordable ? desiredQty : maxAffordable;
-      if (matchQty <= 0) {
-        if (pricePerUnit > 0 && desiredQty > 0) bid.treasuryTruncated = true;
-        return 0;
-      }
-      if (matchQty < desiredQty && pricePerUnit > 0) {
-        bid.treasuryTruncated = true;
-      }
-      filledOut.add(
-        FilledDeal(
-          sellerFactionId: offer.factionId,
-          buyerFactionId: bid.factionId,
-          commodityId: commodityId,
-          quantity: matchQty,
-          pricePerUnit: pricePerUnit,
-          isFirstRightOfRefusalMatch: true,
-          sellerOriginTileKey: offer.order.originTileKey,
-        ),
-      );
-      offer.remaining -= matchQty;
-      bid.remaining -= matchQty;
-      remainingCargo[bid.factionId] = cargoLeft - matchQty;
-      _decrementTreasury(
-        bid: bid,
-        matchQty: matchQty,
-        pricePerUnit: pricePerUnit,
-        remainingTreasury: remainingTreasury,
-      );
-      return matchQty;
-    }
-
-    for (final offer in commodityOffers) {
-      if (offer.remaining <= 0) continue;
-      final originTileKey = offer.order.originTileKey;
-      if (originTileKey == null) continue;
-      final attribution = purchasedTileIndex.attributionForTileKey(
-        originTileKey,
-      );
-      if (attribution == null) continue;
-      final owningGpId = attribution.owningGpId;
-      // FRR overlays the owning GP's purchase intent on the seller's tile;
-      // a buyer == seller scenario would mean the GP somehow auto-offered
-      // its own land, which the index already filters out (purchased tiles
-      // currently owned by a GP are excluded). Guard defensively anyway.
-      if (owningGpId == offer.factionId) continue;
-      for (final bid in commodityBids) {
-        if (offer.remaining <= 0) break;
-        if (bid.factionId != owningGpId) continue;
-        if (bid.remaining <= 0) continue;
-        filledQuantity += attemptFrrMatch(offer, bid);
-      }
-    }
-
-    return filledQuantity;
-  }
-
-  /// Within a priority tier, match lock-recovery sellers (treasury below the
-  /// regiment-build band) before affluent GPs so buyer cargo is not exhausted
-  /// on early faction ids alone (Refs #2924 F12).
-  static void _sortOffersLockRecoverySellersFirst(
-    List<_OrderState> tierOffers, {
-    required Set<String> lockRecoverySellerPriorityIds,
-    required Map<String, int> treasuryByFactionId,
-  }) {
-    if (lockRecoverySellerPriorityIds.isEmpty) return;
-    tierOffers.sort((a, b) {
-      final aPriority = lockRecoverySellerPriorityIds.contains(a.factionId);
-      final bPriority = lockRecoverySellerPriorityIds.contains(b.factionId);
-      if (aPriority != bPriority) return aPriority ? -1 : 1;
-      if (aPriority && bPriority) {
-        final aTreasury = treasuryByFactionId[a.factionId] ?? 0;
-        final bTreasury = treasuryByFactionId[b.factionId] ?? 0;
-        if (aTreasury != bTreasury) return aTreasury.compareTo(bTreasury);
-      }
-      final byFaction = a.factionId.compareTo(b.factionId);
-      if (byFaction != 0) return byFaction;
-      return a.factionLocalIndex.compareTo(b.factionLocalIndex);
-    });
-  }
-
-  static int _runTierMatching({
-    required List<_OrderState> tierOffers,
-    required List<_OrderState> tierBids,
-    required CommodityId commodityId,
-    required double pricePerUnit,
-    required Set<String> ftpPairKeys,
-    required Map<String, int> remainingCargo,
-    required Map<String, int> remainingTreasury,
-    required List<FilledDeal> filledOut,
-  }) {
-    if (tierOffers.isEmpty || tierBids.isEmpty) return 0;
-    var filledQuantity = 0;
-
-    bool ftpEligible(_OrderState offer, _OrderState bid) =>
-        ftpPairKeys.contains(pairKey(offer.factionId, bid.factionId));
-
-    int attemptMatch(_OrderState offer, _OrderState bid, {required bool ftp}) {
-      if (offer.remaining <= 0 || bid.remaining <= 0) return 0;
-      final cargoLeft = remainingCargo[bid.factionId] ?? 0;
-      if (cargoLeft <= 0) return 0;
-      final desiredQty = _min3(offer.remaining, bid.remaining, cargoLeft);
-      if (desiredQty <= 0) return 0;
-      final maxAffordable = _maxAffordableQuantity(
-        bid: bid,
-        pricePerUnit: pricePerUnit,
-        remainingTreasury: remainingTreasury,
-      );
-      final matchQty = desiredQty <= maxAffordable ? desiredQty : maxAffordable;
-      if (matchQty <= 0) {
-        if (pricePerUnit > 0 && desiredQty > 0) bid.treasuryTruncated = true;
-        return 0;
-      }
-      if (matchQty < desiredQty && pricePerUnit > 0) {
-        bid.treasuryTruncated = true;
-      }
-      filledOut.add(
-        FilledDeal(
-          sellerFactionId: offer.factionId,
-          buyerFactionId: bid.factionId,
-          commodityId: commodityId,
-          quantity: matchQty,
-          pricePerUnit: pricePerUnit,
-          isFtpMatch: ftp,
-          sellerOriginTileKey: offer.order.originTileKey,
-        ),
-      );
-      offer.remaining -= matchQty;
-      bid.remaining -= matchQty;
-      remainingCargo[bid.factionId] = cargoLeft - matchQty;
-      _decrementTreasury(
-        bid: bid,
-        matchQty: matchQty,
-        pricePerUnit: pricePerUnit,
-        remainingTreasury: remainingTreasury,
-      );
-      return matchQty;
-    }
-
-    // Pass 1: FTP-eligible matches only.
-    for (final offer in tierOffers) {
-      if (offer.remaining <= 0) continue;
-      for (final bid in tierBids) {
-        if (offer.remaining <= 0) break;
-        if (bid.remaining <= 0) continue;
-        if (!ftpEligible(offer, bid)) continue;
-        filledQuantity += attemptMatch(offer, bid, ftp: true);
-      }
-    }
-
-    // Pass 2: any remaining matches.
-    for (final offer in tierOffers) {
-      if (offer.remaining <= 0) continue;
-      for (final bid in tierBids) {
-        if (offer.remaining <= 0) break;
-        if (bid.remaining <= 0) continue;
-        filledQuantity += attemptMatch(
-          offer,
-          bid,
-          ftp: ftpEligible(offer, bid),
-        );
-      }
-    }
-
-    return filledQuantity;
-  }
-
-  static List<CommodityId> _collectCommodityIds(
-    Map<String, List<TradeOrder>> offers,
-    Map<String, List<TradeOrder>> bids,
-  ) {
-    final set = <CommodityId>{};
-    for (final list in offers.values) {
-      for (final order in list) {
-        set.add(order.commodityId);
-      }
-    }
-    for (final list in bids.values) {
-      for (final order in list) {
-        set.add(order.commodityId);
-      }
-    }
-    final sorted = set.toList()..sort();
-    return sorted;
-  }
-
-  static Map<String, List<_OrderState>> _indexOrdersByFaction(
-    Map<String, List<TradeOrder>> ordersByFaction,
-  ) {
-    final sortedFactionIds = ordersByFaction.keys.toList()..sort();
-    final result = <String, List<_OrderState>>{};
-    for (final factionId in sortedFactionIds) {
-      final orders = ordersByFaction[factionId] ?? const <TradeOrder>[];
-      final states = <_OrderState>[];
-      for (var i = 0; i < orders.length; i++) {
-        states.add(
-          _OrderState(
-            factionId: factionId,
-            order: orders[i],
-            factionLocalIndex: i,
-          ),
-        );
-      }
-      result[factionId] = states;
-    }
-    return result;
-  }
-
-  static List<_OrderState> _orderedStatesForCommodity(
-    Map<String, List<_OrderState>> statesByFaction,
-    CommodityId commodityId,
-  ) {
-    final out = <_OrderState>[];
-    for (final entry in statesByFaction.entries) {
-      for (final state in entry.value) {
-        if (state.order.commodityId == commodityId) {
-          out.add(state);
-        }
-      }
-    }
-    return out;
-  }
-
-  static List<int> _collectPriorityTiers(
-    List<_OrderState> offers,
-    List<_OrderState> bids,
-  ) {
-    final tiers = <int>{};
-    for (final state in offers) {
-      tiers.add(state.order.priority);
-    }
-    for (final state in bids) {
-      tiers.add(state.order.priority);
-    }
-    final sorted = tiers.toList()..sort();
-    return sorted;
-  }
-
-  static int _sumInputQuantity(List<_OrderState> states) {
-    var total = 0;
-    for (final state in states) {
-      total += state.order.quantity;
-    }
-    return total;
-  }
-
-  static Map<String, List<TradeOrder>> _carryForwardByFaction(
-    Map<String, List<_OrderState>> statesByFaction,
-  ) {
-    final result = <String, List<TradeOrder>>{};
-    for (final entry in statesByFaction.entries) {
-      final remaining = <TradeOrder>[];
-      for (final state in entry.value) {
-        if (state.remaining > 0) {
-          remaining.add(state.asCarryForward());
-        }
-      }
-      if (remaining.isNotEmpty) {
-        result[entry.key] = List.unmodifiable(remaining);
-      }
-    }
-    return Map.unmodifiable(result);
-  }
-
-  static int _min3(int a, int b, int c) {
-    var m = a < b ? a : b;
-    if (c < m) m = c;
-    return m;
-  }
-
-  /// Returns the maximum units the buyer can afford to purchase at
-  /// [pricePerUnit] given their remaining treasury budget. The
-  /// `pricePerUnit == 0` branch preserves legacy free-fill behavior on
-  /// the missing-price defect path per
-  /// `SPEC/program/world-market-resolution.md` § Step C (Refs #3115):
-  /// returns a high cap (`bid.remaining`) so the other three clamps
-  /// dominate.
-  static int _maxAffordableQuantity({
-    required _OrderState bid,
-    required double pricePerUnit,
-    required Map<String, int> remainingTreasury,
-  }) {
-    if (pricePerUnit <= 0.0) return bid.remaining;
-    final treasuryLeft = remainingTreasury[bid.factionId] ?? 0;
-    if (treasuryLeft <= 0) return 0;
-    final affordable = (treasuryLeft / pricePerUnit).floor();
-    return affordable < 0 ? 0 : affordable;
-  }
-
-  /// Decrements the per-buyer running treasury tally after a successful
-  /// match. Skips the decrement on the missing-price defect path so
-  /// free-fill behavior is preserved (no treasury accounting when no
-  /// notional is owed).
-  static void _decrementTreasury({
-    required _OrderState bid,
-    required int matchQty,
-    required double pricePerUnit,
-    required Map<String, int> remainingTreasury,
-  }) {
-    if (pricePerUnit <= 0.0 || matchQty <= 0) return;
-    final notional = (matchQty * pricePerUnit).round();
-    final treasuryLeft = remainingTreasury[bid.factionId] ?? 0;
-    final next = treasuryLeft - notional;
-    remainingTreasury[bid.factionId] = next < 0 ? 0 : next;
   }
 }
